Simulate multivariate functional data

Simulate random coefficients and create a multivariate functional data object of class `mfd`. It is mainly for internal use, to check that the package functions work.

data_sim_mfd(nobs = 5, nbasis = 5, nvar = 2, seed)

Arguments

nobs

Number of functional observations to be simulated.

nbasis

Number of basis functions.

nvar

Number of functional covariates.

seed

Deprecated: use set.seed() before calling the function for reproducibility.

Value

A simulated object of class `mfd`.

Examples

library(funcharts)
data_sim_mfd()
#> $coefs
#> , , var1
#> 
#>            rep1       rep2       rep3       rep4       rep5
#> [1,] -2.7163443 -0.2653342  2.2982260  1.1346424 -0.1243974
#> [2,] -0.1623528  1.6556775 -0.2049600  0.1713311  1.5929562
#> [3,]  0.4107938 -0.4068828 -0.7383479 -0.6708879  0.9696353
#> [4,]  1.6864199  1.5904254  1.1926553 -0.3223524  1.8476703
#> [5,]  0.5788437 -0.3236311 -0.9375456 -1.4127611  0.6401604
#> 
#> , , var2
#> 
#>            rep1       rep2        rep3       rep4        rep5
#> [1,] -0.1387155 -1.1683511 -0.87395731 -0.9408162  2.05861202
#> [2,]  0.2970397 -0.6164579 -0.02441779  0.1083836  0.35719684
#> [3,] -1.0385692 -0.5274389  0.80038399  0.2217351 -0.03013733
#> [4,] -1.4275881 -1.7245700 -0.99212498  0.6420766  1.87119889
#> [5,] -0.6633314  0.3899491  1.10489558  1.2021144  0.32714477
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 5
#> 
#> $params
#> [1] 0.5
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4" "bspl4.5"
#> 
#> $B
#>              [,1]         [,2]        [,3]         [,4]         [,5]
#> [1,] 0.0714285714 0.0437500000 0.008928571 0.0008928571 0.0000000000
#> [2,] 0.0437500000 0.1107142857 0.069642857 0.0250000000 0.0008928571
#> [3,] 0.0089285714 0.0696428571 0.092857143 0.0696428571 0.0089285714
#> [4,] 0.0008928571 0.0250000000 0.069642857 0.1107142857 0.0437500000
#> [5,] 0.0000000000 0.0008928571 0.008928571 0.0437500000 0.0714285714
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "t"
#> 
#> $fdnames[[2]]
#> [1] "rep1" "rep2" "rep3" "rep4" "rep5"
#> 
#> $fdnames[[3]]
#> [1] "var1" "var2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd"