Get out of control observations from control charts

Get out of control observations from control charts

get_ooc(cclist)

Arguments

cclist
A data.frame produced by control_charts_pca, control_charts_sof_pc, regr_cc_fof, or regr_cc_sof.

Value

A data.frame with the same number of rows as cclist, and the same number of columns apart from the columns indicating control chart limits. Each value is TRUE if the corresponding observation is in control and FALSE otherwise.

Examples

library(funcharts)
data("air")
air <- lapply(air, function(x) x[1:220, , drop = FALSE])
fun_covariates <- c("CO", "temperature")
mfdobj_x <- get_mfd_list(air[fun_covariates],
                         n_basis = 15,
                         lambda = 1e-2)
mfdobj_y <- get_mfd_list(air["NO2"],
                         n_basis = 15,
                         lambda = 1e-2)
y_scalar <- rowMeans(air$NO2)
y1_scalar <- y_scalar[1:200]
y2_scalar <- y_scalar[201:220]
mfdobj_y1 <- mfdobj_y[1:200]
mfdobj_y2 <- mfdobj_y[201:220]
mfdobj_x1 <- mfdobj_x[1:200]
mfdobj_x2 <- mfdobj_x[201:220]
mod_fof <- fof_pc(mfdobj_y1, mfdobj_x1)
cclist <- regr_cc_fof(mod_fof,
                      mfdobj_y_new = mfdobj_y2,
                      mfdobj_x_new = mfdobj_x2)
get_ooc(cclist)

#>     id    T2   spe
#> 1  201 FALSE FALSE
#> 2  202 FALSE FALSE
#> 3  203 FALSE FALSE
#> 4  204 FALSE FALSE
#> 5  205 FALSE FALSE
#> 6  206 FALSE FALSE
#> 7  207 FALSE FALSE
#> 8  208 FALSE FALSE
#> 9  209 FALSE FALSE
#> 10 210 FALSE  TRUE
#> 11 211 FALSE FALSE
#> 12 212 FALSE FALSE
#> 13 213 FALSE  TRUE
#> 14 214 FALSE  TRUE
#> 15 215 FALSE FALSE
#> 16 216 FALSE FALSE
#> 17 217 FALSE FALSE
#> 18 218 FALSE FALSE
#> 19 219 FALSE FALSE
#> 20 220 FALSE FALSE