Use a function-on-function linear regression model for prediction

Predict new observations of the functional response variable and calculate the corresponding prediction error (and their standardized or studentized version) given new observations of functional covariates and a fitted function-on-function linear regression model.

predict_fof_pc(object, mfdobj_y_new, mfdobj_x_new)

Arguments

object
A list obtained as output from fof_pc, i.e. a fitted function-on-function linear regression model.

mfdobj_y_new
An object of class mfd containing new observations of the functional response.

mfdobj_x_new
An object of class mfd containing new observations of the functional covariates.

Value

A list of mfd objects. It contains:

* pred_error: the prediction error of the standardized functional response variable,

* pred_error_original_scale: the prediction error of the functional response variable on the original scale,

* y_hat_new: the prediction of the functional response observations on the original scale,

* y_z_new: the standardized version of the functional response observations provided in mfdobj_y_new,

* y_hat_z_new: the prediction of the functional response observations on the standardized/studentized scale.

References

Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020) Functional Regression Control Chart. Technometrics. <doi:10.1080/00401706.2020.1753581>

Examples

library(funcharts)
data("air")
air <- lapply(air, function(x) x[1:10, , drop = FALSE])
fun_covariates <- c("CO", "temperature")
mfdobj_x <- get_mfd_list(air[fun_covariates], lambda = 1e-2)
mfdobj_y <- get_mfd_list(air["NO2"], lambda = 1e-2)
mod <- fof_pc(mfdobj_y, mfdobj_x)
predict_fof_pc(mod,
               mfdobj_y_new = mfdobj_y,
               mfdobj_x_new = mfdobj_x)

#> $pred_error
#> $coefs
#> , , NO2
#> 
#>                     1          2            3          4             5
#> bspl4.1  -0.316992896 0.01645679  0.186326845 0.19078829 -0.2022617283
#> bspl4.2  -0.320150264 0.01896032  0.174959476 0.18487340 -0.1842840727
#> bspl4.3  -0.327397470 0.02135022  0.150251682 0.17388711 -0.1436794469
#> bspl4.4  -0.335363955 0.02877685  0.110213513 0.15024024 -0.0847943729
#> bspl4.5  -0.346832967 0.03955042  0.070117260 0.12286427 -0.0357112198
#> bspl4.6  -0.363438024 0.05718032  0.030914925 0.09333851 -0.0040770808
#> bspl4.7  -0.384551721 0.08215908 -0.007971204 0.06650798  0.0073833227
#> bspl4.8  -0.408214134 0.11214152 -0.045481841 0.04621413  0.0001358468
#> bspl4.9  -0.432737618 0.14330604 -0.078006737 0.03285337 -0.0215241814
#> bspl4.10 -0.455975870 0.17318154 -0.104271188 0.02683548 -0.0535458727
#> bspl4.11 -0.477411900 0.19953940 -0.125701106 0.02779014 -0.0880295112
#> bspl4.12 -0.494532713 0.22106644 -0.143773879 0.03556352 -0.1211366571
#> bspl4.13 -0.506419716 0.23720248 -0.160215571 0.05006401 -0.1485155241
#> bspl4.14 -0.510612540 0.24831858 -0.177621741 0.07201515 -0.1666583900
#> bspl4.15 -0.502089560 0.25453804 -0.199799935 0.10419388 -0.1734446381
#> bspl4.16 -0.476070222 0.25504453 -0.231902083 0.14792520 -0.1648210955
#> bspl4.17 -0.428483966 0.24933339 -0.275902508 0.20335536 -0.1368361145
#> bspl4.18 -0.357858722 0.23769107 -0.331605974 0.26951694 -0.0870439912
#> bspl4.19 -0.265995980 0.21958054 -0.397752115 0.34106319 -0.0129889339
#> bspl4.20 -0.154813669 0.19462581 -0.473697912 0.41169476  0.0853041282
#> bspl4.21 -0.027912959 0.16372954 -0.559364476 0.47464186  0.2063870911
#> bspl4.22  0.104356482 0.13240388 -0.652736561 0.52072882  0.3452570224
#> bspl4.23  0.223035140 0.11094329 -0.747414115 0.54236049  0.4819527252
#> bspl4.24  0.298569637 0.11308116 -0.830034953 0.53073799  0.5892754248
#> bspl4.25  0.305093231 0.15063767 -0.879867218 0.47908519  0.6403500011
#> bspl4.26  0.238569422 0.21694709 -0.878008066 0.38859798  0.6220231522
#> bspl4.27  0.126783670 0.29205905 -0.828769591 0.27719597  0.5503488587
#> bspl4.28  0.007958313 0.35887584 -0.756067044 0.16994692  0.4577191464
#> bspl4.29 -0.057877507 0.39081563 -0.706949138 0.11090150  0.3971012587
#> bspl4.30 -0.088212131 0.40492113 -0.683054216 0.08397818  0.3700036334
#>                     6            7            8          9           10
#> bspl4.1   0.262639834 -0.523435353  0.019745031 0.34637775  0.020355437
#> bspl4.2   0.242034101 -0.518220375  0.008248248 0.36181500  0.031764172
#> bspl4.3   0.198368766 -0.504081898 -0.014922651 0.39140150  0.054822183
#> bspl4.4   0.130355763 -0.470582296 -0.049256129 0.43330502  0.087105357
#> bspl4.5   0.067883779 -0.422667498 -0.077543870 0.47238289  0.109956936
#> bspl4.6   0.015727812 -0.361685194 -0.096607363 0.51034814  0.118297954
#> bspl4.7  -0.024107376 -0.291428736 -0.105636445 0.54771996  0.109925132
#> bspl4.8  -0.050827513 -0.217558487 -0.106411264 0.58416893  0.085832811
#> bspl4.9  -0.064241302 -0.146113012 -0.102507626 0.61917302  0.049798048
#> bspl4.10 -0.065518779 -0.082817532 -0.098876425 0.65226116  0.008727491
#> bspl4.11 -0.053650461 -0.032707283 -0.100580640 0.68281510 -0.032063739
#> bspl4.12 -0.030572581  0.002396862 -0.112352076 0.71097641 -0.067635330
#> bspl4.13  0.001313467  0.022253032 -0.136338940 0.73772949 -0.097072718
#> bspl4.14  0.035871421  0.029226337 -0.171338167 0.76509680 -0.124297457
#> bspl4.15  0.063307826  0.028725809 -0.213523494 0.79422709 -0.156135017
#> bspl4.16  0.077121545  0.026083779 -0.258501757 0.82330443 -0.198184328
#> bspl4.17  0.069590809  0.025036775 -0.303491761 0.85059729 -0.253199276
#> bspl4.18  0.033225948  0.027913702 -0.346348501 0.87394824 -0.319438717
#> bspl4.19 -0.034957790  0.038346468 -0.386386838 0.88959745 -0.390505992
#> bspl4.20 -0.133651340  0.062563029 -0.425552331 0.89031431 -0.456786781
#> bspl4.21 -0.258167871  0.106472074 -0.466968559 0.86772701 -0.506543703
#> bspl4.22 -0.400102443  0.172662601 -0.512408900 0.81649238 -0.526653280
#> bspl4.23 -0.543723297  0.261937927 -0.559268443 0.73441217 -0.504235884
#> bspl4.24 -0.662943449  0.362773560 -0.603662287 0.62926097 -0.427058059
#> bspl4.25 -0.731571764  0.453315714 -0.635855161 0.50914180 -0.290329466
#> bspl4.26 -0.733906116  0.512645980 -0.643805372 0.38443451 -0.107498574
#> bspl4.27 -0.684228695  0.533050791 -0.621091466 0.26800041  0.086650999
#> bspl4.28 -0.612969125  0.525172779 -0.575565561 0.16838506  0.256543675
#> bspl4.29 -0.564340457  0.511310290 -0.541016318 0.11554038  0.344514367
#> bspl4.30 -0.541346857  0.504608996 -0.524426859 0.09124481  0.382283323
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames[[3]]
#> [1] "NO2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd" 
#> 
#> $pred_error_original_scale
#> $coefs
#> , , NO2
#> 
#>                     1           2             3           4             5
#> bspl4.1  -0.022007533 0.001142734  0.0129360305 0.013245545 -1.404250e-02
#> bspl4.2  -0.021935882 0.001300131  0.0119750324 0.012660441 -1.260670e-02
#> bspl4.3  -0.021946785 0.001433901  0.0100480439 0.011642275 -9.595460e-03
#> bspl4.4  -0.022124866 0.001900165  0.0072647008 0.009906862 -5.586589e-03
#> bspl4.5  -0.022938699 0.002612879  0.0046485748 0.008134601 -2.374350e-03
#> bspl4.6  -0.024537547 0.003848794  0.0021144585 0.006324862 -2.917354e-04
#> bspl4.7  -0.026959142 0.005735107 -0.0005178435 0.004692664  5.153913e-04
#> bspl4.8  -0.030127631 0.008242346 -0.0033103732 0.003438618  3.024553e-05
#> bspl4.9  -0.033892887 0.011187447 -0.0060698959 0.002591641 -1.643201e-03
#> bspl4.10 -0.037902197 0.014364171 -0.0086385717 0.002236389 -4.400659e-03
#> bspl4.11 -0.041785773 0.017444340 -0.0109835823 0.002427612 -7.664886e-03
#> bspl4.12 -0.044869771 0.020048444 -0.0130349561 0.003217996 -1.096979e-02
#> bspl4.13 -0.046645933 0.021847700 -0.0147554676 0.004608357 -1.367830e-02
#> bspl4.14 -0.046710270 0.022720588 -0.0162575455 0.006600459 -1.525348e-02
#> bspl4.15 -0.044735107 0.022690716 -0.0178341137 0.009324506 -1.546006e-02
#> bspl4.16 -0.040677414 0.021814281 -0.0198882182 0.012720062 -1.407653e-02
#> bspl4.17 -0.034717559 0.020240986 -0.0224879860 0.016611949 -1.105769e-02
#> bspl4.18 -0.027260525 0.018169659 -0.0254791465 0.020741224 -6.564314e-03
#> bspl4.19 -0.018915354 0.015715000 -0.0286430612 0.024578584 -7.926646e-04
#> bspl4.20 -0.010185201 0.012994245 -0.0318583751 0.027685247  5.884394e-03
#> bspl4.21 -0.001556797 0.010190782 -0.0351039078 0.029757024  1.309358e-02
#> bspl4.22  0.006266821 0.007717023 -0.0383292573 0.030526480  2.038559e-02
#> bspl4.23  0.012433332 0.006129496 -0.0414437561 0.030017297  2.678420e-02
#> bspl4.24  0.015908538 0.006033187 -0.0442195674 0.028237246  3.140555e-02
#> bspl4.25  0.016096756 0.007943997 -0.0464229437 0.025291469  3.378103e-02
#> bspl4.26  0.013096049 0.011757093 -0.0478954153 0.021290690  3.395852e-02
#> bspl4.27  0.007678749 0.017033231 -0.0488439950 0.016507041  3.251504e-02
#> bspl4.28  0.000808651 0.023333032 -0.0496702385 0.011377211  3.019533e-02
#> bspl4.29 -0.004027027 0.027750744 -0.0503424295 0.007966356  2.832340e-02
#> bspl4.30 -0.006539844 0.030024133 -0.0506489037 0.006227578  2.743549e-02
#>                      6             7             8           9           10
#> bspl4.1   0.0182341720 -0.0363402393  0.0013708045 0.024047784  0.001413203
#> bspl4.2   0.0165614489 -0.0354962888  0.0005544828 0.024800667  0.002186666
#> bspl4.3   0.0132569745 -0.0337675261 -0.0010194952 0.026255146  0.003692925
#> bspl4.4   0.0085899738 -0.0310367040 -0.0032537192 0.028590342  0.005749834
#> bspl4.5   0.0045065140 -0.0279713228 -0.0051220585 0.031236198  0.007267663
#> bspl4.6   0.0010942935 -0.0244758690 -0.0065157207 0.034448684  0.007989781
#> bspl4.7  -0.0016547337 -0.0205267280 -0.0074068938 0.038392551  0.007729627
#> bspl4.8  -0.0037271981 -0.0161782673 -0.0078656418 0.043112775  0.006385126
#> bspl4.9  -0.0050251130 -0.0115657941 -0.0080451502 0.048495184  0.003967769
#> bspl4.10 -0.0054599204 -0.0069805078 -0.0082295548 0.054218783  0.000792068
#> bspl4.11 -0.0047230432 -0.0029191022 -0.0088001108 0.059762200 -0.002757655
#> bspl4.12 -0.0027975811  0.0001979933 -0.0101808838 0.064502450 -0.006113899
#> bspl4.13  0.0001166108  0.0020499813 -0.0125531819 0.067948679 -0.008938445
#> bspl4.14  0.0032992212  0.0026754896 -0.0156933744 0.070003549 -0.011384637
#> bspl4.15  0.0056658719  0.0025579142 -0.0190760802 0.070818499 -0.013952150
#> bspl4.16  0.0066047045  0.0022280938 -0.0221742542 0.070467117 -0.017017844
#> bspl4.17  0.0056219783  0.0020344364 -0.0247206988 0.069135572 -0.020660983
#> bspl4.18  0.0024600402  0.0021456637 -0.0265759542 0.066925740 -0.024562387
#> bspl4.19 -0.0026422912  0.0027761105 -0.0277765300 0.063822739 -0.028122533
#> bspl4.20 -0.0091279279  0.0042430198 -0.0285693099 0.059627416 -0.030693508
#> bspl4.21 -0.0163354849  0.0067357329 -0.0292600991 0.054203775 -0.031724603
#> bspl4.22 -0.0236016453  0.0102045283 -0.0300546364 0.047716182 -0.030831086
#> bspl4.23 -0.0302115289  0.0145857827 -0.0309914253 0.040553051 -0.027856444
#> bspl4.24 -0.0353348599  0.0193594159 -0.0321561593 0.033446327 -0.022679677
#> bspl4.25 -0.0385923638  0.0239046710 -0.0335498050 0.026892555 -0.015345365
#> bspl4.26 -0.0400376634  0.0279072411 -0.0351067364 0.021085549 -0.006055327
#> bspl4.27 -0.0403455363  0.0313287906 -0.0365800419 0.015960528  0.004746198
#> bspl4.28 -0.0403019790  0.0344083639 -0.0377951937 0.011250114  0.016394709
#> bspl4.29 -0.0402011428  0.0363811426 -0.0385222929 0.008288194  0.024383054
#> bspl4.30 -0.0401408341  0.0374166228 -0.0388864344 0.006766263  0.028345932
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames[[3]]
#> [1] "NO2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd" 
#> 
#> $y_hat_new
#> $coefs
#> , , NO2
#> 
#>                 1        2        3        4        5        6        7
#> bspl4.1  7.149229 7.125976 7.234536 7.340986 7.302005 7.251271 7.253769
#> bspl4.2  7.150800 7.129598 7.236169 7.337782 7.308382 7.255807 7.257492
#> bspl4.3  7.154099 7.137027 7.239442 7.331225 7.320998 7.264841 7.264900
#> bspl4.4  7.159536 7.148383 7.244709 7.321734 7.340708 7.279040 7.276502
#> bspl4.5  7.166544 7.160841 7.250938 7.312800 7.361880 7.294661 7.289221
#> bspl4.6  7.175994 7.175185 7.258792 7.304859 7.385337 7.312520 7.303780
#> bspl4.7  7.188597 7.192158 7.268868 7.298369 7.411575 7.333187 7.320746
#> bspl4.8  7.204750 7.212114 7.281571 7.294006 7.440574 7.356803 7.340353
#> bspl4.9  7.224429 7.234844 7.296924 7.292416 7.471666 7.382958 7.362334
#> bspl4.10 7.246980 7.259444 7.314376 7.293878 7.503523 7.410595 7.385803
#> bspl4.11 7.271351 7.284548 7.332861 7.298260 7.533928 7.437947 7.409208
#> bspl4.12 7.296056 7.308382 7.350912 7.304995 7.560360 7.462898 7.430643
#> bspl4.13 7.320026 7.329787 7.367329 7.313273 7.580693 7.483720 7.448542
#> bspl4.14 7.342769 7.348684 7.381639 7.322220 7.594222 7.499804 7.462399
#> bspl4.15 7.364177 7.365832 7.394037 7.330990 7.601776 7.511664 7.472753
#> bspl4.16 7.384257 7.382142 7.405048 7.339224 7.604670 7.520182 7.480532
#> bspl4.17 7.402785 7.398385 7.415384 7.347100 7.604522 7.526383 7.486886
#> bspl4.18 7.419315 7.414975 7.425703 7.355233 7.602678 7.531053 7.492805
#> bspl4.19 7.433353 7.432186 7.436602 7.364418 7.600154 7.534761 7.499102
#> bspl4.20 7.444458 7.450130 7.448442 7.375078 7.597810 7.537960 7.506370
#> bspl4.21 7.452294 7.468833 7.461272 7.387143 7.595993 7.540780 7.514803
#> bspl4.22 7.456551 7.488048 7.474789 7.400254 7.594361 7.542896 7.524092
#> bspl4.23 7.456750 7.507078 7.488330 7.413756 7.592338 7.543742 7.533586
#> bspl4.24 7.452538 7.524619 7.500839 7.426980 7.588766 7.542374 7.542093
#> bspl4.25 7.443816 7.539709 7.511375 7.438979 7.582816 7.538132 7.548637
#> bspl4.26 7.430966 7.551824 7.519474 7.449260 7.574299 7.530932 7.552775
#> bspl4.27 7.414919 7.561446 7.525466 7.457978 7.563524 7.521248 7.554796
#> bspl4.28 7.397066 7.569747 7.530205 7.465688 7.551143 7.509955 7.555469
#> bspl4.29 7.384653 7.575169 7.533139 7.470534 7.542477 7.502018 7.555661
#> bspl4.30 7.378536 7.577813 7.534576 7.472988 7.538095 7.498037 7.555704
#>                 8        9       10
#> bspl4.1  7.263572 7.199747 7.212254
#> bspl4.2  7.267389 7.203548 7.214530
#> bspl4.3  7.274966 7.211205 7.219125
#> bspl4.4  7.286977 7.223312 7.226529
#> bspl4.5  7.300364 7.237014 7.235282
#> bspl4.6  7.315914 7.253190 7.246169
#> bspl4.7  7.334181 7.272480 7.259815
#> bspl4.8  7.355348 7.295092 7.276552
#> bspl4.9  7.379096 7.320703 7.296274
#> bspl4.10 7.404489 7.348329 7.318270
#> bspl4.11 7.429902 7.376388 7.341302
#> bspl4.12 7.453352 7.402922 7.363733
#> bspl4.13 7.473179 7.426414 7.384261
#> bspl4.14 7.488717 7.446307 7.402330
#> bspl4.15 7.500324 7.462909 7.418046
#> bspl4.16 7.508738 7.476772 7.431780
#> bspl4.17 7.514878 7.488387 7.443924
#> bspl4.18 7.519495 7.497960 7.454716
#> bspl4.19 7.523154 7.505526 7.464303
#> bspl4.20 7.526294 7.511109 7.472746
#> bspl4.21 7.529018 7.514635 7.479956
#> bspl4.22 7.530992 7.515781 7.485613
#> bspl4.23 7.531655 7.514042 7.489165
#> bspl4.24 7.530134 7.508769 7.489917
#> bspl4.25 7.525788 7.499625 7.487376
#> bspl4.26 7.518546 7.486828 7.481562
#> bspl4.27 7.508871 7.471105 7.473093
#> bspl4.28 7.497613 7.453604 7.463034
#> bspl4.29 7.489684 7.441447 7.455915
#> bspl4.30 7.485718 7.435407 7.452385
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames[[3]]
#> [1] "NO2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd" 
#> 
#> $y_z_new
#> $coefs
#> , , NO2
#> 
#>                   1           2           3          4         5           6
#> bspl4.1  -1.5284298 -1.52990871  0.20363021  1.7413738 0.7868508  0.52098738
#> bspl4.2  -1.5655716 -1.53629908  0.17546391  1.6693425 0.8692440  0.52859735
#> bspl4.3  -1.6350481 -1.54173632  0.11577917  1.5104885 1.0367508  0.54171737
#> bspl4.4  -1.7178445 -1.52295083  0.01888392  1.2270229 1.2785939  0.55911534
#> bspl4.5  -1.7742946 -1.47379561 -0.08155151  0.9054650 1.4907202  0.57777855
#> bspl4.6  -1.8020546 -1.39270405 -0.18260342  0.5591866 1.6575813  0.59889303
#> bspl4.7  -1.8006271 -1.28228771 -0.28147934  0.2094467 1.7698591  0.62076124
#> bspl4.8  -1.7735295 -1.15272081 -0.37321681 -0.1178313 1.8261788  0.64134513
#> bspl4.9  -1.7286953 -1.01933000 -0.45177289 -0.4026208 1.8344536  0.66060762
#> bspl4.10 -1.6755072 -0.89638780 -0.51587104 -0.6342901 1.8087988  0.68027879
#> bspl4.11 -1.6206290 -0.79312073 -0.56810016 -0.8115619 1.7656198  0.70440113
#> bspl4.12 -1.5668543 -0.71565221 -0.61243031 -0.9398126 1.7181636  0.73498832
#> bspl4.13 -1.5127139 -0.66317926 -0.65306167 -1.0297192 1.6750618  0.77211227
#> bspl4.14 -1.4526296 -0.62879906 -0.69422805 -1.0938068 1.6405834  0.81084605
#> bspl4.15 -1.3788225 -0.60309106 -0.74047369 -1.1431807 1.6165948  0.84206722
#> bspl4.16 -1.2846610 -0.57765697 -0.79601364 -1.1846379 1.6046290  0.85853744
#> bspl4.17 -1.1692848 -0.54527944 -0.86005826 -1.2199983 1.6075272  0.85154585
#> bspl4.18 -1.0393418 -0.50079389 -0.92839095 -1.2462125 1.6287976  0.81276583
#> bspl4.19 -0.9079293 -0.44011433 -0.99402936 -1.2596490 1.6718779  0.73771360
#> bspl4.20 -0.7901250 -0.35914996 -1.05057911 -1.2582392 1.7395040  0.62680323
#> bspl4.21 -0.7010314 -0.25000334 -1.09172626 -1.2412138 1.8307260  0.48299137
#> bspl4.22 -0.6581960 -0.09891113 -1.10824471 -1.2062096 1.9350697  0.31057262
#> bspl4.23 -0.6829901  0.10742466 -1.08786120 -1.1446798 2.0242483  0.12037958
#> bspl4.24 -0.7922019  0.37203919 -1.01687886 -1.0433836 2.0564807 -0.06729487
#> bspl4.25 -0.9805159  0.68378823 -0.88397181 -0.8968016 1.9889264 -0.22963327
#> bspl4.26 -1.2069966  0.99462817 -0.69440351 -0.7134720 1.8044252 -0.34570834
#> bspl4.27 -1.4114659  1.25280829 -0.48012127 -0.5170635 1.5352011 -0.41573974
#> bspl4.28 -1.5551888  1.43829766 -0.28029200 -0.3338687 1.2415557 -0.45497502
#> bspl4.29 -1.6102447  1.51801380 -0.17050133 -0.2310317 1.0614279 -0.46780473
#> bspl4.30 -1.6308895  1.54971684 -0.12135493 -0.1848988 0.9791493 -0.47241619
#>                     7            8           9         10
#> bspl4.1  -0.229094253  0.455281774 -0.13740830 -0.2832830
#> bspl4.2  -0.206904338  0.463985340 -0.11419306 -0.2836651
#> bspl4.3  -0.159548033  0.479746253 -0.06515076 -0.2829988
#> bspl4.4  -0.080195013  0.499902630  0.01736735 -0.2798956
#> bspl4.5   0.004812897  0.518434711  0.11067667 -0.2782463
#> bspl4.6   0.091610793  0.536439813  0.21521340 -0.2815629
#> bspl4.7   0.175409575  0.552855925  0.32774181 -0.2916802
#> bspl4.8   0.251156690  0.565452953  0.44184243 -0.3086775
#> bspl4.9   0.314944329  0.572539558  0.55096382 -0.3310899
#> bspl4.10  0.364390448  0.573136153  0.65054529 -0.3550933
#> bspl4.11  0.396779446  0.565380124  0.73862099 -0.3773897
#> bspl4.12  0.412314064  0.547937375  0.81600341 -0.3946574
#> bspl4.13  0.411106196  0.520010018  0.88642784 -0.4060442
#> bspl4.14  0.395455744  0.482476498  0.95505440 -0.4149526
#> bspl4.15  0.371232307  0.438072443  1.02552488 -0.4279236
#> bspl4.16  0.344390183  0.389237718  1.09663410 -0.4504589
#> bspl4.17  0.321136228  0.336995871  1.16409697 -0.4866813
#> bspl4.18  0.307892038  0.282394650  1.22047601 -0.5375870
#> bspl4.19  0.313579690  0.224717042  1.25427751 -0.6004437
#> bspl4.20  0.350564486  0.161070133  1.25005886 -0.6699074
#> bspl4.21  0.430486229  0.086415633  1.19144398 -0.7380884
#> bspl4.22  0.559711572 -0.004752509  1.06526060 -0.7943005
#> bspl4.23  0.739832862 -0.113362256  0.86328031 -0.8262724
#> bspl4.24  0.951354360 -0.237862719  0.59459633 -0.8168487
#> bspl4.25  1.155047936 -0.367843489  0.28107662 -0.7500731
#> bspl4.26  1.305062981 -0.482568274 -0.03751029 -0.6234573
#> bspl4.27  1.376909099 -0.562476798 -0.31721086 -0.4608404
#> bspl4.28  1.382756347 -0.605298694 -0.53455841 -0.2984281
#> bspl4.29  1.363180519 -0.617643256 -0.63943506 -0.2059615
#> bspl4.30  1.351240859 -0.621632109 -0.68446188 -0.1644536
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames[[3]]
#> [1] "NO2"
#> 
#> 
#> $id_var
#> [1] "id"
#> 
#> attr(,"class")
#> [1] "mfd" "fd" 
#> attr(,"scaled:center")
#> $coefs
#> , , mean NO2
#> 
#>           mean
#>  [1,] 7.233334
#>  [2,] 7.236150
#>  [3,] 7.241783
#>  [4,] 7.250743
#>  [5,] 7.260955
#>  [6,] 7.273174
#>  [7,] 7.287998
#>  [8,] 7.305716
#>  [9,] 7.326164
#> [10,] 7.348569
#> [11,] 7.371570
#> [12,] 7.393425
#> [13,] 7.412722
#> [14,] 7.428909
#> [15,] 7.442251
#> [16,] 7.453334
#> [17,] 7.462863
#> [18,] 7.471393
#> [19,] 7.479356
#> [20,] 7.487040
#> [21,] 7.494473
#> [22,] 7.501338
#> [23,] 7.507044
#> [24,] 7.510703
#> [25,] 7.511625
#> [26,] 7.509647
#> [27,] 7.505245
#> [28,] 7.499352
#> [29,] 7.495070
#> [30,] 7.492926
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#> [1] "mean"
#> 
#> $fdnames[[3]]
#> [1] "mean NO2"
#> 
#> 
#> attr(,"class")
#> [1] "fd"
#> attr(,"scaled:scale")
#> $coefs
#>             [,1]
#>  [1,] 0.06942641
#>  [2,] 0.06850617
#>  [3,] 0.06701949
#>  [4,] 0.06596880
#>  [5,] 0.06614638
#>  [6,] 0.06755004
#>  [7,] 0.07016752
#>  [8,] 0.07388782
#>  [9,] 0.07841031
#> [10,] 0.08320123
#> [11,] 0.08757894
#> [12,] 0.09075331
#> [13,] 0.09210948
#> [14,] 0.09147721
#> [15,] 0.08912861
#> [16,] 0.08554402
#> [17,] 0.08123270
#> [18,] 0.07654115
#> [19,] 0.07172565
#> [20,] 0.06699075
#> [21,] 0.06252645
#> [22,] 0.05854246
#> [23,] 0.05533865
#> [24,] 0.05323887
#> [25,] 0.05277533
#> [26,] 0.05452709
#> [27,] 0.05879251
#> [28,] 0.06543809
#> [29,] 0.07110952
#> [30,] 0.07414986
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames$<NA>
#> [1] "arg"
#> 
#> $fdnames$`Std. Dev.`
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames$`Std. Dev. NA`
#> [1] "NO2"
#> 
#> 
#> attr(,"class")
#> [1] "fd"
#> 
#> $y_hat_z_new
#> $coefs
#> , , NO2
#> 
#>                   1            2             3          4         5          6
#> bspl4.1  -1.2114369 -1.546365501  0.0173033695  1.5505855 0.9891126 0.25834755
#> bspl4.2  -1.2454213 -1.555259394  0.0005044313  1.4844691 1.0535281 0.28656325
#> bspl4.3  -1.3076507 -1.563086539 -0.0344725137  1.3366014 1.1804302 0.34334861
#> bspl4.4  -1.3824806 -1.551727683 -0.0913295900  1.0767826 1.3633882 0.42875958
#> bspl4.5  -1.4274617 -1.513346034 -0.1516687718  0.7826007 1.5264315 0.50989477
#> bspl4.6  -1.4386166 -1.449884375 -0.2135183476  0.4658481 1.6616584 0.58316522
#> bspl4.7  -1.4160754 -1.364446791 -0.2735081392  0.1429387 1.7624758 0.64486861
#> bspl4.8  -1.3653154 -1.264862328 -0.3277349730 -0.1640454 1.8260429 0.69217265
#> bspl4.9  -1.2959577 -1.162636035 -0.3737661510 -0.4354742 1.8559778 0.72484893
#> bspl4.10 -1.2195313 -1.069569339 -0.4115998541 -0.6611256 1.8623447 0.74579757
#> bspl4.11 -1.1432171 -0.992660125 -0.4423990581 -0.8393520 1.8536493 0.75805160
#> bspl4.12 -1.0723215 -0.936718652 -0.4686564256 -0.9753761 1.8393002 0.76556090
#> bspl4.13 -1.0062942 -0.900381733 -0.4928461024 -1.0797832 1.8235773 0.77079880
#> bspl4.14 -0.9420170 -0.877117644 -0.5166063101 -1.1658220 1.8072418 0.77497463
#> bspl4.15 -0.8767330 -0.857629098 -0.5406737502 -1.2473746 1.7900394 0.77875940
#> bspl4.16 -0.8085908 -0.832701503 -0.5641115542 -1.3325631 1.7694501 0.78141590
#> bspl4.17 -0.7408009 -0.794612831 -0.5841557545 -1.4233537 1.7443634 0.78195504
#> bspl4.18 -0.6814831 -0.738484967 -0.5967849797 -1.5157294 1.7158416 0.77953988
#> bspl4.19 -0.6419333 -0.659694878 -0.5962772420 -1.6007122 1.6848668 0.77267139
#> bspl4.20 -0.6353114 -0.553775767 -0.5768811991 -1.6699339 1.6541998 0.76045457
#> bspl4.21 -0.6731184 -0.413732874 -0.5323617861 -1.7158557 1.6243390 0.74115924
#> bspl4.22 -0.7625525 -0.231315005 -0.4555081487 -1.7269385 1.5898127 0.71067506
#> bspl4.23 -0.9060252 -0.003518628 -0.3404470842 -1.6870403 1.5422955 0.66410288
#> bspl4.24 -1.0907715  0.258958027 -0.1868439046 -1.5741216 1.4672053 0.59564858
#> bspl4.25 -1.2856091  0.533150560 -0.0041045887 -1.3758868 1.3485764 0.50193849
#> bspl4.26 -1.4455660  0.777681081  0.1836045607 -1.1020700 1.1824020 0.38819777
#> bspl4.27 -1.5382496  0.960749232  0.3486483177 -0.7942595 0.9848523 0.26848896
#> bspl4.28 -1.5631471  1.079421819  0.4757750479 -0.5038156 0.7838366 0.15799410
#> bspl4.29 -1.5523672  1.127198172  0.5364478107 -0.3419332 0.6643267 0.09653572
#> bspl4.30 -1.5426773  1.144795710  0.5616992837 -0.2688770 0.6091457 0.06893066
#>                  7           8            9         10
#> bspl4.1  0.2943411  0.43553674 -0.483786046 -0.3036384
#> bspl4.2  0.3113160  0.45573709 -0.476008056 -0.3154293
#> bspl4.3  0.3445339  0.49466890 -0.456552256 -0.3378210
#> bspl4.4  0.3903873  0.54915876 -0.415937672 -0.3670009
#> bspl4.5  0.4274804  0.59597858 -0.361706216 -0.3882032
#> bspl4.6  0.4532960  0.63304718 -0.295134734 -0.3998608
#> bspl4.7  0.4668383  0.65849237 -0.219978156 -0.4016053
#> bspl4.8  0.4687152  0.67186422 -0.142326499 -0.3945103
#> bspl4.9  0.4610573  0.67504718 -0.068209194 -0.3808880
#> bspl4.10 0.4472080  0.67201258 -0.001715866 -0.3638208
#> bspl4.11 0.4294867  0.66596076  0.055805885 -0.3453260
#> bspl4.12 0.4099172  0.66028945  0.105027007 -0.3270221
#> bspl4.13 0.3888532  0.65634896  0.148698351 -0.3089715
#> bspl4.14 0.3662294  0.65381467  0.189957598 -0.2906551
#> bspl4.15 0.3425065  0.65159594  0.231297790 -0.2717886
#> bspl4.16 0.3183064  0.64773947  0.273329667 -0.2522746
#> bspl4.17 0.2960995  0.64048763  0.313499687 -0.2334820
#> bspl4.18 0.2799783  0.62874315  0.346527770 -0.2181483
#> bspl4.19 0.2752332  0.61110388  0.364680066 -0.2099377
#> bspl4.20 0.2880015  0.58662246  0.359744554 -0.2131206
#> bspl4.21 0.3240142  0.55338419  0.323716972 -0.2315447
#> bspl4.22 0.3870490  0.50765639  0.248768214 -0.2676472
#> bspl4.23 0.4778949  0.44590619  0.128868143 -0.3220365
#> bspl4.24 0.5885808  0.36579957 -0.034664640 -0.3897906
#> bspl4.25 0.7017322  0.26801167 -0.228065178 -0.4597436
#> bspl4.26 0.7924170  0.16123710 -0.421944797 -0.5159587
#> bspl4.27 0.8438583  0.05861467 -0.585211267 -0.5474914
#> bspl4.28 0.8575836 -0.02973313 -0.702943470 -0.5549718
#> bspl4.29 0.8518702 -0.07662694 -0.754975441 -0.5504758
#> bspl4.30 0.8466319 -0.09720525 -0.775706688 -0.5467369
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 30
#> 
#> $params
#>  [1] 0.03703704 0.07407407 0.11111111 0.14814815 0.18518519 0.22222222
#>  [7] 0.25925926 0.29629630 0.33333333 0.37037037 0.40740741 0.44444444
#> [13] 0.48148148 0.51851852 0.55555556 0.59259259 0.62962963 0.66666667
#> [19] 0.70370370 0.74074074 0.77777778 0.81481481 0.85185185 0.88888889
#> [25] 0.92592593 0.96296296
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#>  [1] "bspl4.1"  "bspl4.2"  "bspl4.3"  "bspl4.4"  "bspl4.5"  "bspl4.6" 
#>  [7] "bspl4.7"  "bspl4.8"  "bspl4.9"  "bspl4.10" "bspl4.11" "bspl4.12"
#> [13] "bspl4.13" "bspl4.14" "bspl4.15" "bspl4.16" "bspl4.17" "bspl4.18"
#> [19] "bspl4.19" "bspl4.20" "bspl4.21" "bspl4.22" "bspl4.23" "bspl4.24"
#> [25] "bspl4.25" "bspl4.26" "bspl4.27" "bspl4.28" "bspl4.29" "bspl4.30"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "arg"
#> 
#> $fdnames[[2]]
#>  [1] "1"  "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9"  "10"
#> 
#> $fdnames[[3]]
#> [1] "NO2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd" 
#>