Use a scalar-on-function linear regression model for prediction

Predict new observations of the scalar response variable and calculate the corresponding prediction error, with prediction interval limits, given new observations of functional covariates and a fitted scalar-on-function linear regression model

predict_sof_pc(object, newdata = NULL, alpha = 0.05)

Arguments

object
A list obtained as output from sof_pc, i.e. a fitted scalar-on-function linear regression model.

newdata
An object of class mfd containing new observations of the functional covariates. If NULL, it is set as the functional covariates data used for model fitting.

alpha
A numeric value indicating the Type I error for the regression control chart and such that this function returns the 1-alpha prediction interval on the response. Default is 0.05.

Value

A data.frame with as many rows as the number of functional replications in newdata, with the following columns:

* fit: the predictions of the response variable corresponding to new_data,

* lwr: lower limit of the 1-alpha prediction interval on the response,

* upr: upper limit of the 1-alpha prediction interval on the response.

Examples

library(funcharts)
data("air")
air <- lapply(air, function(x) x[1:10, , drop = FALSE])
fun_covariates <- names(air)[names(air) != "NO2"]
mfdobj_x <- get_mfd_list(air[fun_covariates], lambda = 1e-2)
y <- rowMeans(air$NO2)
mod <- sof_pc(y, mfdobj_x)
predict_sof_pc(mod)

#>         fit      lwr      upr
#> 1  7.306757 7.239627 7.373888
#> 2  7.363669 7.296200 7.431139
#> 3  7.391444 7.327498 7.455390
#> 4  7.359601 7.293523 7.425680
#> 5  7.498679 7.436206 7.561152
#> 6  7.456403 7.399907 7.512900
#> 7  7.439771 7.376226 7.503315
#> 8  7.442382 7.384245 7.500519
#> 9  7.453164 7.388864 7.517465
#> 10 7.378384 7.317635 7.439132