Simulate example data for funcharts

Function used to simulate three data sets to illustrate the use of funcharts. It uses the function simulate_mfd, which creates a data set with three functional covariates, a functional response generated as a function of the three functional covariates, and a scalar response generated as a function of the three functional covariates. This function generates three data sets, one for phase I, one for tuning, i.e., to estimate the control chart limits, and one for phase II monitoring. see also simulate_mfd.

sim_funcharts(nobs1 = 1000, nobs_tun = 1000, nobs2 = 60)

Arguments

nobs1

The number of observation to simulate in phase I. Default is 1000.

nobs_tun

The number of observation to simulate the tuning data set. Default is 1000.

nobs2

The number of observation to simulate in phase II. Default is 60.

Value

A list with three objects, datI contains the phase I data, datI_tun contains the tuning data, datII contains the phase II data. In the phase II data, the first group of observations are in control, the second group of observations contains a moderate mean shift, while the third group of observations contains a severe mean shift. The shift types are described in the paper from Centofanti et al. (2021),

References

Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2021) Functional Regression Control Chart. Technometrics, 63(3), 281--294. <doi:10.1080/00401706.2020.1753581>