Simulate example data for funcharts

Function used to simulate a data set to illustrate the use of funcharts. It creates a data set with three functional covariates, a functional response generated as a function of the three functional covariates through a function-on-function linear model, and a scalar response generated as a function of the three functional covariates through a scalar-on-function linear model. This function covers the simulation study in Centofanti et al. (2020) for the function-on-function case and also simulates data in a similar way for the scalar response case. In the default case, the function generates in-control data. Additional arguments can be used to generate additional data that are out of control, with mean shifts according to the scenarios proposed by Centofanti et al. (2020). Each simulated observation of a functional variable consists of a vector of 150 discrete points, equally spaced between 0 and 1, generated with noise.

simulate_mfd(
  nobs = 1000,
  R2 = 0.97,
  seed = 0,
  shift_type_y = "0",
  shift_type_x1 = "0",
  shift_type_x2 = "0",
  shift_type_x3 = "0",
  d_y = 0,
  d_x1 = 0,
  d_x2 = 0,
  d_x3 = 0,
  d_y_scalar = 0,
  save_beta = FALSE
)

Arguments

nobs
The number of observation to simulate

R2
The desired coefficient of determination in the regression. In both the scalar and functional response cases, only three values are allowed, i.e. 0.97, 0.86, 0.74. Default is 0.97.

seed
Set seed for reproducibility. Default is 0.

shift_type_y
The shift type for the functional response. There are five possibilities: "0" if there is no shift, "A", "B", "C" or "D" for the corresponding shift types shown in Centofanti et al. (2020). Default is "0".

shift_type_x1
#' The shift type for the first functional covariate. There are five possibilities: "0" if there is no shift, "A", "B", "C" or "D" for the corresponding shift types shown in Centofanti et al. (2020). Default is "0".

shift_type_x2
#' The shift type for the second functional covariate. #' There are five possibilities: "0" if there is no shift, "A", "B", "C" or "D" for the corresponding shift types shown in Centofanti et al. (2020). Default is "0".

shift_type_x3
#' The shift type for the third functional covariate. #' There are five possibilities: "0" if there is no shift, "A", "B", "C" or "D" for the corresponding shift types shown in Centofanti et al. (2020). Default is "0".

d_y
A number indicating the severity of the shift type for the functional response. Default is 0.

d_x1
A number indicating the severity of the shift type for the first functional covariate. Default is 0.

d_x2
A number indicating the severity of the shift type for the second functional covariate. Default is 0.

d_x3
A number indicating the severity of the shift type for the third functional covariate. Default is 0.

d_y_scalar
A number indicating the severity of the shift type for the scalar response. Default is 0.

save_beta
If TRUE, the true regression coefficients of both the function-on-function and the scalar-on-function models are saved. Default is FALSE.

Value

A list with the following elements:

* X1 is a nobsx150 matrix with the simulated observations of the first functional covariate

* X2 is a nobsx150 matrix with the simulated observations of the second functional covariate

* X3 is a nobsx150 matrix with the simulated observations of the third functional covariate

* Y is a nobsx150 matrix with the simulated observations of the functional response

* y_scalar is a vector of length 150 with the simulated observations of the scalar response

* beta_fof, if save_beta = TRUE, is a list of three 500x500 matrices with the discretized functional coefficients of the funciton-on-function regression

* beta_sof, if save_beta = TRUE, is a list of three vectors of length 500 with the discretized functional coefficients of the scalar-on-function regression

References

Centofanti F, Lepore A, Menafoglio A, Palumbo B, Vantini S. (2020) Functional Regression Control Chart. Technometrics. <doi:10.1080/00401706.2020.1753581>