Tensor product of two Multivariate Functional Data objects

This function returns the tensor product of two Multivariate Functional Data objects. Each object must contain only one replication.

tensor_product_mfd(mfdobj1, mfdobj2 = NULL)

Arguments

mfdobj1
A multivariate functional data object, of class mfd, having only one functional observation.

mfdobj2
A multivariate functional data object, of class mfd, having only one functional observation. If NULL, it is set equal to mfdobj1. Default is NULL.

Value

An object of class bifd. If we denote with x(s)=(x_1(s),...,x_p(s)) the vector of p functions represented by mfdobj1 and with y(t)=(y_1(t),...,y_q(t)) the vector of q functions represented by mfdobj2, the output is the vector of pq bivariate functions

f(s,t)=(x_1(s)y_1(t),...,x_1(s)y_q(t), ...,x_p(s)y_1(t),...,x_p(s)y_q(t)).

Examples

library(funcharts)
mfdobj1 <- data_sim_mfd(nobs = 1, nvar = 3)
mfdobj2 <- data_sim_mfd(nobs = 1, nvar = 2)
tensor_product_mfd(mfdobj1)

#> $coefs
#> , , s.rep1 t.rep1, var1 var1
#> 
#>            bspl4.1    bspl4.2    bspl4.3    bspl4.4
#> bspl4.1  1.5950535 -0.4120178  1.6794757  1.6070201
#> bspl4.2 -0.4120178  0.1064282 -0.4338249 -0.4151089
#> bspl4.3  1.6794757 -0.4338249  1.7683661  1.6920756
#> bspl4.4  1.6070201 -0.4151089  1.6920756  1.6190764
#> 
#> , , s.rep1 t.rep1, var2 var1
#> 
#>            bspl4.1    bspl4.2    bspl4.3     bspl4.4
#> bspl4.1  0.5236732 -1.9448865 -1.1727377 -0.37221845
#> bspl4.2 -0.1352699  0.5023831  0.3029295  0.09614764
#> bspl4.3  0.5513899 -2.0478244 -1.2348078 -0.39191903
#> bspl4.4  0.5276019 -1.9594776 -1.1815359 -0.37501094
#> 
#> , , s.rep1 t.rep1, var3 var1
#> 
#>              bspl4.1    bspl4.2    bspl4.3    bspl4.4
#> bspl4.1 -0.007283676  3.0369673  0.9643836 -1.0091122
#> bspl4.2  0.001881444 -0.7844782 -0.2491097  0.2606635
#> bspl4.3 -0.007669182  3.1977063  1.0154260 -1.0625219
#> bspl4.4 -0.007338320  3.0597515  0.9716187 -1.0166828
#> 
#> , , s.rep1 t.rep1, var1 var2
#> 
#>            bspl4.1     bspl4.2    bspl4.3    bspl4.4
#> bspl4.1  0.5236732 -0.13526987  0.5513899  0.5276019
#> bspl4.2 -1.9448865  0.50238308 -2.0478244 -1.9594776
#> bspl4.3 -1.1727377  0.30292954 -1.2348078 -1.1815359
#> bspl4.4 -0.3722185  0.09614764 -0.3919190 -0.3750109
#> 
#> , , s.rep1 t.rep1, var2 var2
#> 
#>            bspl4.1    bspl4.2    bspl4.3     bspl4.4
#> bspl4.1  0.1719275 -0.6385271 -0.3850224 -0.12220331
#> bspl4.2 -0.6385271  2.3714461  1.4299468  0.45385476
#> bspl4.3 -0.3850224  1.4299468  0.8622367  0.27366769
#> bspl4.4 -0.1222033  0.4538548  0.2736677  0.08686014
#> 
#> , , s.rep1 t.rep1, var3 var2
#> 
#>              bspl4.1    bspl4.2    bspl4.3    bspl4.4
#> bspl4.1 -0.002391309  0.9970689  0.3166175 -0.3313023
#> bspl4.2  0.008881158 -3.7030461 -1.1758958  1.2304343
#> bspl4.3  0.005355206 -2.2328819 -0.7090477  0.7419336
#> bspl4.4  0.001699704 -0.7087005 -0.2250466  0.2354844
#> 
#> , , s.rep1 t.rep1, var1 var3
#> 
#>              bspl4.1      bspl4.2      bspl4.3     bspl4.4
#> bspl4.1 -0.007283676  0.001881444 -0.007669182 -0.00733832
#> bspl4.2  3.036967301 -0.784478156  3.197706304  3.05975148
#> bspl4.3  0.964383634 -0.249109661  1.015426022  0.97161871
#> bspl4.4 -1.009112155  0.260663473 -1.062521910 -1.01668280
#> 
#> , , s.rep1 t.rep1, var2 var3
#> 
#>              bspl4.1      bspl4.2      bspl4.3      bspl4.4
#> bspl4.1 -0.002391309  0.008881158  0.005355206  0.001699704
#> bspl4.2  0.997068931 -3.703046087 -2.232881867 -0.708700521
#> bspl4.3  0.316617488 -1.175895782 -0.709047716 -0.225046606
#> bspl4.4 -0.331302341  1.230434326  0.741933649  0.235484363
#> 
#> , , s.rep1 t.rep1, var3 var3
#> 
#>               bspl4.1     bspl4.2      bspl4.3      bspl4.4
#> bspl4.1  3.326028e-05 -0.01386805 -0.004403775  0.004608024
#> bspl4.2 -1.386805e-02  5.78235792  1.836177604 -1.921340298
#> bspl4.3 -4.403775e-03  1.83617760  0.583074974 -0.610118238
#> bspl4.4  4.608024e-03 -1.92134030 -0.610118238  0.638415780
#> 
#> 
#> $sbasis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 4
#> 
#> $params
#> NULL
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $tbasis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 4
#> 
#> $params
#> NULL
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $bifdnames
#> $bifdnames[[1]]
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> $bifdnames[[2]]
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> $bifdnames[[3]]
#> [1] "s.rep1 t.rep1"
#> 
#> $bifdnames[[4]]
#> [1] "var1 var1" "var2 var1" "var3 var1" "var1 var2" "var2 var2" "var3 var2"
#> [7] "var1 var3" "var2 var3" "var3 var3"
#> 
#> 
#> attr(,"class")
#> [1] "bifd"

tensor_product_mfd(mfdobj1, mfdobj2)

#> $coefs
#> , , s.rep1 t.rep1, var1 var1
#> 
#>            bspl4.1    bspl4.2    bspl4.3    bspl4.4
#> bspl4.1  1.5950535 -0.4120178  1.6794757  1.6070201
#> bspl4.2 -0.4120178  0.1064282 -0.4338249 -0.4151089
#> bspl4.3  1.6794757 -0.4338249  1.7683661  1.6920756
#> bspl4.4  1.6070201 -0.4151089  1.6920756  1.6190764
#> 
#> , , s.rep1 t.rep1, var2 var1
#> 
#>            bspl4.1    bspl4.2    bspl4.3     bspl4.4
#> bspl4.1  0.5236732 -1.9448865 -1.1727377 -0.37221845
#> bspl4.2 -0.1352699  0.5023831  0.3029295  0.09614764
#> bspl4.3  0.5513899 -2.0478244 -1.2348078 -0.39191903
#> bspl4.4  0.5276019 -1.9594776 -1.1815359 -0.37501094
#> 
#> , , s.rep1 t.rep1, var3 var1
#> 
#>            bspl4.1     bspl4.2    bspl4.3    bspl4.4
#> bspl4.1  0.5236732 -0.13526987  0.5513899  0.5276019
#> bspl4.2 -1.9448865  0.50238308 -2.0478244 -1.9594776
#> bspl4.3 -1.1727377  0.30292954 -1.2348078 -1.1815359
#> bspl4.4 -0.3722185  0.09614764 -0.3919190 -0.3750109
#> 
#> , , s.rep1 t.rep1, var1 var2
#> 
#>            bspl4.1    bspl4.2    bspl4.3     bspl4.4
#> bspl4.1  0.1719275 -0.6385271 -0.3850224 -0.12220331
#> bspl4.2 -0.6385271  2.3714461  1.4299468  0.45385476
#> bspl4.3 -0.3850224  1.4299468  0.8622367  0.27366769
#> bspl4.4 -0.1222033  0.4538548  0.2736677  0.08686014
#> 
#> , , s.rep1 t.rep1, var2 var2
#> 
#>              bspl4.1      bspl4.2      bspl4.3     bspl4.4
#> bspl4.1 -0.007283676  0.001881444 -0.007669182 -0.00733832
#> bspl4.2  3.036967301 -0.784478156  3.197706304  3.05975148
#> bspl4.3  0.964383634 -0.249109661  1.015426022  0.97161871
#> bspl4.4 -1.009112155  0.260663473 -1.062521910 -1.01668280
#> 
#> , , s.rep1 t.rep1, var3 var2
#> 
#>              bspl4.1      bspl4.2      bspl4.3      bspl4.4
#> bspl4.1 -0.002391309  0.008881158  0.005355206  0.001699704
#> bspl4.2  0.997068931 -3.703046087 -2.232881867 -0.708700521
#> bspl4.3  0.316617488 -1.175895782 -0.709047716 -0.225046606
#> bspl4.4 -0.331302341  1.230434326  0.741933649  0.235484363
#> 
#> 
#> $sbasis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 4
#> 
#> $params
#> NULL
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $tbasis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 4
#> 
#> $params
#> NULL
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $bifdnames
#> $bifdnames[[1]]
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> $bifdnames[[2]]
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4"
#> 
#> $bifdnames[[3]]
#> [1] "s.rep1 t.rep1"
#> 
#> $bifdnames[[4]]
#> [1] "var1 var1" "var2 var1" "var3 var1" "var1 var2" "var2 var2" "var3 var2"
#> 
#> 
#> attr(,"class")
#> [1] "bifd"