Simulate multivariate functional data

Simulate random coefficients and create a multivariate functional data object of class `mfd`. It is mainly for internal use, to check that the package functions work.

data_sim_mfd(nobs = 5, nbasis = 5, nvar = 2, seed)

Arguments

nobs

Number of functional observations to be simulated.

nbasis

Number of basis functions.

nvar

Number of functional covariates.

seed

Deprecated: use set.seed() before calling the function for reproducibility.

Value

A simulated object of class `mfd`.

Examples

library(funcharts)
data_sim_mfd()
#> $coefs
#> , , var1
#> 
#>            rep1       rep2       rep3       rep4       rep5
#> [1,] -0.4498284  1.9882214  0.2722130  1.2916589  0.6835232
#> [2,]  0.4586420  1.4499486 -0.5728849 -0.8628613 -0.8738104
#> [3,] -0.9319680  1.1128207  1.9378192  2.6633988  0.1257748
#> [4,] -0.7019924 -0.8996759 -1.7377162  0.3801997 -0.6454739
#> [5,] -0.5391790 -0.3207785 -0.6349188 -0.5049375  0.6308045
#> 
#> , , var2
#> 
#>             rep1       rep2        rep3        rep4       rep5
#> [1,]  0.08921166 -1.2996241 -0.90417988 -0.82874106  0.4382774
#> [2,]  0.84930293 -1.1408272  0.02461763 -0.22937701  0.2063968
#> [3,] -0.63139895 -0.8568484  1.40971049  0.58246939 -0.3592978
#> [4,] -0.45039936 -0.5397717  0.40873667 -0.04541096 -0.6709019
#> [5,] -0.93450739  1.3156163 -0.69381547 -0.76420763  0.6626958
#> 
#> 
#> $basis
#> $call
#> basisfd(type = type, rangeval = rangeval, nbasis = nbasis, params = params, 
#>     dropind = dropind, quadvals = quadvals, values = values, 
#>     basisvalues = basisvalues)
#> 
#> $type
#> [1] "bspline"
#> 
#> $rangeval
#> [1] 0 1
#> 
#> $nbasis
#> [1] 5
#> 
#> $params
#> [1] 0.5
#> 
#> $dropind
#> NULL
#> 
#> $quadvals
#> NULL
#> 
#> $values
#> list()
#> 
#> $basisvalues
#> list()
#> 
#> $names
#> [1] "bspl4.1" "bspl4.2" "bspl4.3" "bspl4.4" "bspl4.5"
#> 
#> $B
#>              [,1]         [,2]        [,3]         [,4]         [,5]
#> [1,] 0.0714285714 0.0437500000 0.008928571 0.0008928571 0.0000000000
#> [2,] 0.0437500000 0.1107142857 0.069642857 0.0250000000 0.0008928571
#> [3,] 0.0089285714 0.0696428571 0.092857143 0.0696428571 0.0089285714
#> [4,] 0.0008928571 0.0250000000 0.069642857 0.1107142857 0.0437500000
#> [5,] 0.0000000000 0.0008928571 0.008928571 0.0437500000 0.0714285714
#> 
#> attr(,"class")
#> [1] "basisfd"
#> 
#> $fdnames
#> $fdnames[[1]]
#> [1] "t"
#> 
#> $fdnames[[2]]
#> [1] "rep1" "rep2" "rep3" "rep4" "rep5"
#> 
#> $fdnames[[3]]
#> [1] "var1" "var2"
#> 
#> 
#> attr(,"class")
#> [1] "mfd" "fd"