Get out of control observations from control charts

Get out of control observations from control charts

get_ooc(cclist)

Arguments

cclist

A data.frame produced by control_charts_pca, control_charts_sof_pc, regr_cc_fof, or regr_cc_sof.

Value

A data.frame with the same number of rows as cclist, and the same number of columns apart from the columns indicating control chart limits. Each value is TRUE if the corresponding observation is in control and FALSE otherwise.

Examples

library(funcharts)
data("air")
air <- lapply(air, function(x) x[201:300, , drop = FALSE])
fun_covariates <- c("CO", "temperature")
mfdobj_x <- get_mfd_list(air[fun_covariates],
                         n_basis = 15,
                         lambda = 1e-2)
y <- rowMeans(air$NO2)
y1 <- y[1:60]
y2 <- y[91:100]
mfdobj_x1 <- mfdobj_x[1:60]
mfdobj_x_tuning <- mfdobj_x[61:90]
mfdobj_x2 <- mfdobj_x[91:100]
mod <- sof_pc(y1, mfdobj_x1)
cclist <- control_charts_sof_pc(mod = mod,
                                y_test = y2,
                                mfdobj_x_test = mfdobj_x2,
                                mfdobj_x_tuning = mfdobj_x_tuning)
get_ooc(cclist)
#>      id    T2   spe     y contribution_T2_CO contribution_T2_temperature
#> 91   91 FALSE FALSE FALSE              FALSE                       FALSE
#> 92   92 FALSE FALSE FALSE              FALSE                       FALSE
#> 93   93  TRUE FALSE FALSE               TRUE                       FALSE
#> 94   94 FALSE FALSE FALSE              FALSE                       FALSE
#> 95   95 FALSE FALSE FALSE              FALSE                       FALSE
#> 96   96 FALSE FALSE FALSE              FALSE                       FALSE
#> 97   97 FALSE FALSE FALSE              FALSE                       FALSE
#> 98   98 FALSE FALSE FALSE              FALSE                       FALSE
#> 99   99 FALSE FALSE  TRUE              FALSE                       FALSE
#> 100 100 FALSE FALSE  TRUE              FALSE                       FALSE
#>     contribution_spe_CO contribution_spe_temperature
#> 91                FALSE                        FALSE
#> 92                FALSE                        FALSE
#> 93                FALSE                        FALSE
#> 94                FALSE                        FALSE
#> 95                FALSE                        FALSE
#> 96                FALSE                        FALSE
#> 97                FALSE                        FALSE
#> 98                FALSE                        FALSE
#> 99                FALSE                        FALSE
#> 100               FALSE                        FALSE